backend/requirements.txt
django
djangorestframework
django-cors-headers
pip install djangorestframework-simplejwt